Some expansive changes here.
Most work is on the editor. It has been pulled from the Pages system and now is on it's own subroot. It will use actual paths and communicate using a SharedWorker. This system is now in place and am now building up (current work on getting the graphics going on the main [terrain] editor).

Also some other minor changes. THREE upgraded from 67 to 69. WEBIX added (serves as core UI framework of editor).