Textures & Lighting
This commit introduces full fledged texturing with lighting and fog effects for terrain and other possible meshes. This is done using custom shaders, tile maps, and texture maps. the conversion is not entirely complete: tileMaps are to be stored using uint8 and generated into textures for the shaders, but right now the uint8 values are ignored and static images are used (see temp folder). This will change in the future.

All changes basically reflect this change. Some modifications on Mod standard: tileMap replaced by bottomTileMap and topTileMap, which is somewhat obvious. The bottomTileMap is placed first, then the top, then an "infoMap" which is made up of the pathingMap and dynamic values (i.e., when placing structures or modifying terrain). The tiles variable was replaced with tileTextures, which link an image and describes said image (horizontal/vertical tiles and size).