Pathing & Info Map
General changes reflect moving towards having the info map and pathing map work. Basically, the info map is a combination of a pathing map and an "active" map, which is used to indicate live changes (such as when modifying terrain, placing a structure, etc).

Mod
-tileMapInfo > tileMapPathing; size reduced to 2 because the texture is always 0
-saved property, handles emitting on changes (so saving in on window modifies others)

Terrain HTML/CSS/JS
-Added styling & logic for toggling, using showPathingMap as the first one of its kind
-Incorporated logic for using emitted save/unsave events

CanvasMerge
-New class being introduced to merge two canvases
-Simple creates a third canvas that will be used as the texture for the plane material
-Idea is to modify the top (active) canvas, then recalc to push changes