Texturing and stuff...
Got very basic texturing working. Next goal is to get brush sizes working (easy for height transformer, more complicated for texture transformer). The texture transformer is going to require a vast amount of rules to define the circumstance to use a variation. Further, it needs to push top layer textures down to the bottom one. getImageData is not required for this because it should be stored in the uint8's anyways.